# Optimize x86/aarch64 MD5 implementation