### Description of changes: 

(Equivalent to https://github.com/openssl/openssl/commit/ebe34f9a62630b45a825bc07a2e9cf52731e836e)

As suggested in https://github.com/animetosho/md5-optimisation?tab=readme-ov-file#dependency-shortcut-in-g-function, we can delay the dependency on 'x' by recognizing that ((x & z) | (y & ~z)) is equivalent to ((x & z) + (y + ~z)) in this scenario, and we can perform those additions independently, leaving our dependency on x to the final addition. This speeds it up around 5% on both platforms.

### Call-outs:
I applied the x86 patch manually, it's trivial, and the aarch64 patch applied cleanly after fixing the file name/path.

### Testing:
`ninja-build run_tests` on x86. `bssl speed -filter MD5` on x86 shows the expected speedup:

```
Before
Did 7169250 MD5 (16 bytes) operations in 1000024us (7169077.9 ops/sec): 114.7 MB/s
Did 1826500 MD5 (256 bytes) operations in 1000119us (1826282.7 ops/sec): 467.5 MB/s
Did 433000 MD5 (1350 bytes) operations in 1001295us (432440.0 ops/sec): 583.8 MB/s
Did 76000 MD5 (8192 bytes) operations in 1005792us (75562.3 ops/sec): 619.0 MB/s
Did 39000 MD5 (16384 bytes) operations in 1025281us (38038.4 ops/sec): 623.2 MB/s


After
Did 7457250 MD5 (16 bytes) operations in 1000004us (7457220.2 ops/sec): 119.3 MB/s
Did 1919000 MD5 (256 bytes) operations in 1000057us (1918890.6 ops/sec): 491.2 MB/s
Did 456000 MD5 (1350 bytes) operations in 1001377us (455373.0 ops/sec): 614.8 MB/s
Did 80000 MD5 (8192 bytes) operations in 1004118us (79671.9 ops/sec): 652.7 MB/s
Did 41000 MD5 (16384 bytes) operations in 1022055us (40115.3 ops/sec): 657.2 MB/s 
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and the ISC license.